cleaning up badge container + adding new props
