+++ src/pages/Badges/index.tsx
@@ -1,7 +1,19 @@
 import React from 'react';
+import BadgeContainer from '../../components/Badges/BadgeContainer';
 
 const Badges = (): React.ReactElement => {
-  return <p>placeholder badge page</p>;
+  return (
+    <>
+      <BadgeContainer
+        badgeId={''}
+        image={''}
+        isEarned={false}
+        onClick={function (): void {
+          throw new Error('Function not implemented.');
+        }}
+      />
+    </>
+  );
 };
 
 export default Badges;
